Change to the admin layout
An attempt to make the admin form more user-friendly by grouping the fields based on elements visible in the front end, rather than on a flow that might be lesss intuitive for a novice. Also changed all fieldset variables to $fs so the fields code could be moved around more easily.