Changed form layout and fields definitions syntax
Changed the configuration form layout a bit, to a layout that feels easier on the eyes to me.
The fields definitions were changed to the array syntax (also because it feels easier on the eyes to me :smile: ). I commented out the toolbar definitions for the CKEditor fields to see if using the defaults would make issues appear and it seems ok.