Update PrivacyWire.js
Added data-poster to allowed elements to update data-poster attribut. 

Explanation:
I use PrivacyWire with the InstagramMediaDisplay module, among others, to initially block files and only load them when the user agrees. 
Unfortunately, I have not found a way to block the poster attribute in a <video> element. I hope this is possible with this change?

Could you please check this and adopt it if necessary?